update：将config.py中的部分代码分离出来；
