+++ config.py
@@ -1,29 +1,13 @@
 import json
-import os
 
-
-# 方法区
-
-# 获取项目路径
-def get_work_dir():
-    return os.path.dirname(os.path.abspath(__file__))
-
-
-# 获取项目的缓存路径
-def get_cache_dir():
-    return get_work_dir() + '/cache'
-
-
-# 根据相对路径生成绝对路径
-def get_abs_path(*rel_paths):
-    return os.path.join(get_work_dir(), *rel_paths)
+from common import PathUtils
 
 
 def init_config():
     print('配置初始化中...')
-    work_dir = get_work_dir()
+    work_dir = PathUtils.get_work_dir()
     print('脚本工作路径: ' + work_dir)
-    cache_dir = get_cache_dir()
+    cache_dir = PathUtils.get_cache_dir()
     print('缓存文件路径: ' + cache_dir)
     print()
 
@@ -36,6 +20,6 @@ def init_config():
 
 
 def get(key):
-    with open(get_cache_dir() + '/runtime.json') as json_file:
+    with open(PathUtils.get_cache_dir() + '/runtime.json') as json_file:
         data = json.load(json_file)
         return data[key]
